requirements.txt
django==4.1.3
djangorestframework==3.14.0
pyyaml==6.0.1
uritemplate==4.1.1
psycopg2-binary
pgvector
sentence-transformers
sumy
fake_useragent
lxml
selenium
selenium-wire
undetected_chromedriver